+++ src/evaluators/expr.h
@@ -12,4 +12,4 @@ struct mud_expr_evaluator_s {
 mud_expr_evaluator_t * mud_expr_evaluator_init(mud_expr_t * expr, mud_scope_t * scope);
 void mud_expr_evaluator_free(mud_expr_evaluator_t * evaluator);
 
-mud_object_t * _mud_expr_evaluator_get(mud_expr_evaluator_t * evaluator, unsigned i);
+mud_object_t * _mud_expr_evaluator_get(mud_expr_evaluator_t * evaluator, size_t i);
